category card 컴포넌트 분리
